invites/requirements.txt
pyjwt
websockets